Twitter Sentiment Analysis Buildup 1
This repo contains the python codes that are done to build a sentiment
analyzer for tweets.

Note: The python module "twitter_app_login" - is a private module - it
contains the security credentials that are obtained from twitter to
stream the tweets. To get your please visit https://apps.twitter.com/